+++ tango/integrations/flax/model.py
@@ -6,9 +6,9 @@ from tango.common.registrable import Registrable
 class Model(nn.Module, Registrable):
     """
     This is a :class:`~tango.common.Registrable` mixin class that inherits from
-    :class:`flax.linen.nn.Module`.
-    Its :meth:`~flax.linen.nn.Module.setup()` can be used to register submodules,
+    :class:`flax.linen.Module`.
+    Its :meth:`~flax.linen.Module.setup()` can be used to register submodules,
     variables, parameters you will need in your model.
-    Its :meth:`~flax.linen.nn.Module.__call__()` returns the output of the model
+    Its :meth:`~flax.linen.Module.__call__()` returns the output of the model
     for a given input.
     """